Extract parts of VariantHeader that are specific to FocusPage
