+++ src/pages/FocusPage.tsx
@@ -6,6 +6,7 @@ import { VariantAgeDistributionPlotWidget } from '../widgets/VariantAgeDistribut
 import { VariantTimeDistributionPlotWidget } from '../widgets/VariantTimeDistributionPlot';
 import { NamedSection } from '../components/NamedSection';
 import Switzerland from '../components/Switzerland';
+import { FocusVariantHeaderControls } from '../components/FocusVariantHeaderControls';
 
 interface Props {
   country: Country;
@@ -14,14 +15,19 @@ interface Props {
 }
 
 export const FocusPage = (props: Props) => {
+  const { country, matchPercentage, variant } = props;
   const plotProps = {
-    country: props.country,
-    matchPercentage: props.matchPercentage,
-    mutations: props.variant.mutations,
+    country,
+    matchPercentage,
+    mutations: variant.mutations,
   };
   return (
     <>
-      <VariantHeader {...props} />
+      <VariantHeader variant={variant} controls={<FocusVariantHeaderControls {...props} />} />
+      <p style={{ marginBottom: '30px' }}>
+        The following plots show sequences matching <b>{Math.round(matchPercentage * 100)}%</b> of the
+        mutations.
+      </p>
       <NamedSection title='Sequences over time'>
         <VariantTimeDistributionPlotWidget.ShareableComponent {...plotProps} height={300} />
       </NamedSection>
